move dream workflow to separate file
